+++ flake.nix
@@ -13,7 +13,7 @@
 
         repoctx = pkgs.rustPlatform.buildRustPackage {
           pname = "repoctx";
-          version = "0.8.0";
+          version = "0.9.0";
           src = ./.;
 
           cargoLock = {